requirements.txt
discord.py
python-dotenv~=0.14.0
nanoid~=2.0.0